Implement a grid-stride loop for better memory coalescing.
This is really important for performance on A100s. Unfortunately, it comes at
the cost of our output index compression scheme, so we'll have to get creative
if we want to bring that back. My naive approaches to reorder the input and
output indices to allow for delta compression resulted in very poor read
performance, possibly due to bank conflicts, so reimplementing delta compression
efficiently may be a long term project.